Add an inheritance example.
The parent/child chart relationship is roughly:
inheritance -> secure-elasticsearch -> elasticsearch

The objective is to demonstrate a parent chart (secure-elasticsearch)
which sets broad business-specific defaults for Elasticsearch, and this
chart is then used to deploy master and data nodes, inheriting those
settings.

This example covers a few areas:

1) Having a child chart (secure-elasticsearch) add some baseline
   settings for the cluster.
2) A parent chart to inherit and amend the settings of the child
   secure-elasticsearch chart.
3) Using chart aliases to deploy master and data nodes independently
   Same result as multi in PR #22, but instead of using yaml anchors,
   this uses helm child charts.
4) Add a job which deploy the license payload.
5) Add a job for deploying cluster and index settings. This can be
   useful for setting cluster-wide things like concurrent_recoveries, or
   for index-level settings such as node_left recovery delay. These
   settings live along-side the node settings intended for easier reading.
6) (From previous commit) Elasticsearch node settings
   (elasticsearch.yml) is now specified by a structured value instead of
   a text block.

The `Makefile` does a few things to help make this testable outside of
Elastic's infrastructure. Mainly, all secrets and key material are
generated by `make secrets`. This is instead of using the `vault read`
method that other examples are using.

Note: The `secure-elasticsearch` chart is in a parent directory because
for some reason helm doesn't work if it were in
./charts/secure-elasticsearch due to some behavior in helm itself:
https://github.com/helm/helm/issues/2887