Only require setting masterService for custom master nodeGroups
This change means that the when the masterService is not set it will
always point to $clusterName-master-headless. Now masterService only
needs to be set when the nodeGroup name of the master is something
different. For most users this just means one less value to set when
sticking with the default naming for the master nodes group.