Add migration guide for migrating from the community helm/charts repo
Related: https://github.com/helm/charts/issues/10543

This is the first step in making it possible for users of the
helm/charts version to migrate to this chart. The other
part of this is writing up a guide to help users convert their existing
values files into something compatible with the new chart.