[ci] add JJBB job definitions
Add JJBB job definitions to the repo so they can be managed here. Update
job config to use new auto-generated `inject-passwords` section. Note
that the `infra-root+build@elastic.co` is now the default, so it'll be
appended to the list of existing publishers on any jobs that have them,
like the `trigger-parameterized-builds` publisher for
`elastic+helm-charts+pull-request`.