WIP: Add metricbeat chart
Right now this is literally the filebeat chart with a
s/filebeat/metricbeat run over it and the default metricbeat.yml
changed. Since it was already working in this state I'm opening up a PR
so that it is easier to see the changes made versus filebeat and already
get CI testing it.