[kibana] Always set server.host to the docker default
Fixes: #156

The Kibana docker image comes with
[server.host](https://github.com/elastic/kibana/blob/3a1d4ad0acd986f55b7da532878e842dd62a1237/src/dev/build/tasks/os_packages/docker_generator/templates/kibana_yml.template.js#L30)
set in the default `kibana.yml`. This setting is needed to make sure
that services and ingress work correctly. If someone overrides the
kibana.yml with custom configuration this setting changes back to the
kibana default which is `localhost`.

This keeps the setting consistent for the default helm install with
default values and for anyone customising `kibana.yml`