Remove goss port test and instead do http calls on 0.0.0.0 and service
With the default install on GKE 1.13 the default bound port is now ipv4
instead of ipv6. There is an open issue in goss
https://github.com/aelsabbahy/goss/issues/149 to allow testing for
situations like this where it is listening on both ports.

However the only important thing to test is to make sure that this this
port is listening publicly and that the service actually works.

Also switched the security example to test against the service to make
sure we don't hit the same kibana bug as in #156