Don't mount in the current directory when generating certs
This gets really tricky when you are doing docker in docker because the
host path of the host doesn't match up properly with the host of the
local machine.

This also means that we can remove the "run as the current user id" hack
to make sure that mounted files were accessible by the hosted user.