[metricbeat] Fix default configuration for kubernetes module
This configuration didn't work at all since the nodes localhost is not
reachable from the kubernetes pod. The test has also been fixed up to
make sure that the data coming in actually has proper fields in it. The
test was passing because the events were coming in but just with errors.