Add working examples for running Elasticsearch and Kibana on openshift
Closes: #15

We don't currently have automated testing for openshift clusters however
these have been manually tested and confirmed to work. For now having
an example on how to do this is nice but in the future we can consider
adding a value to pragmatically disable these features with something
like `openshift: true`

Special thanks to @geogdog for getting these working and providing the
correct changes needed.