[kibana] Explicitly test for a 200 for readinessProbe
If authentication wasn't configured correctly the Kibana readinessProbe
would still pass since `curl --fail` doesn't consider a redirect to be a
failure. If authentication fails Kibana would redirect you back to the
login page so this readinessProbe always passed for security clusters.
The tests are going to fail until #272 is merged and this PR has been
rebased on top of it.