[script] Update bumper script to handle SNAPSHOT versions (#581)
* [script] Update bumper script to handle SNAPSHOT versions

This commit updates the `bumper.py` script to handle the
workflows for updating the versions to test staging artifacts
and before a release, for example:

- update 7.7.0-SNAPSHOT to 7.7.0 before tagging a release, by running
  `env BUMPER_VERSION_7="7.7.0" ./helpers/bumper.py`
- update 7.7.0-SNAPHSHOT to 7.7.0-abcdefgh for testing staging artifacts
by running:
  `env BUMPER_VERSION_7="7.7.0-abcdefgh" BUMPER_USE_STAGING_IMAGES="true" ./helpers/bumper.py`

For now this script do not handle the bump from 7.7.0 to 7.7.1-SNAPSHOT
for the next developement iteration after a release. In the meantime
this update should be done manually.