[logstash] remove expected logstash host and name assertion (#724)
This commit remove expected Logstash host and name in Logstash tests with 6.x.

When using a headless service (4a6067e), Logstash 6.x use the full name of the service related to the namespace (ie: helm-logstash-default-logstash-0.helm-logstash-oss-logstash-headless.helm-charts-testing.svc.cluster.local). Keeping this test would require a hard coupling between the test and the namespace where it is running.