[kibana] disable nss dentry cache
This commit disable nss dentry cache for readinessProbe.

On affected system (nss < 3.52), the curl commands used by
readinessProbe are filling dentry cache which is observed in some
cases to be never reclaimed.

Kibana Docker images are based on CentOS 7 which is using
nss-3.44.0-7.el7_7.x86_64 and are affected by this bug.

This commit disable nss dentry cache in readinessProbe as a
workaround.