logstash secret creation with template                                                                                                                    [19/878]
Add the ability to specify a secret value using a template.

When the logstash chart is a dependent chart, this allows the parent chart
to specify a secret's contents by using a value.

For example, the parent chart can be installed with a command such as:

`helm install foo bar --set token=secret123`

rather than having to specify the full structure of the secret.

The logstash chart's values.yaml for the above example would look
similar to:

```
    secrets:
      - name: logz-token
        value:
          LOGZ_TOKEN.template: "{{ .Values.logstash.token }}"
```

Combined with envFrom this creates a simple way for
users of the parent chart to set known environment variables to secret values.

Further details on the use case here:

  https://github.com/elastic/helm-charts/issues/1150