add apm-server labels to pods
APM Server labels are applied at the deployment level only, not at
the pod level. This does not align with the behavior of other Elastic
helm charts in this collection. Additionally, it creates issues when
applying labels for use with validation tools, such as
[OPA Gatekeeper](https://github.com/open-policy-agent/gatekeeper).

Add labels defined under the `labels: {}` tag in `values.yml` to
correct this issue.