fix: add labels at pod level, as well as deployment level
APM Server labels are applied at the deployment level only, not at the pod level. This does not align with the behavior of other Elastic helm charts in this collection. Additionally, it creates issues when using tools, which require specific labels to be applied to pods (i.e., validation tools, such as [OPA Gatekeeper](https://github.com/open-policy-agent/gatekeeper)).

Add labels defined under the `labels: {}` tag in `values.yml` to correct this issue.