[meta] Initiate 7.17 branch
This commit initiates the 7.17 branch which will be dedicated to the 7.17.x releases for the Helm Charts.

It will allow to:

test this branch with the daily Stack Docker images 7.17.x-SNAPSHOT via dedicated Jenkins jobs
test the staging 7.17.x Docker images before a release
This branch is based on 7.x which was initiated in #605.