[elasticsearch] do not regenerate certs when they already exists
This commit update the way we generate the Elasticsearch certs to lookup
if the certs secret already exists and reuse the secrets instead of regenerate
the certs.

This fix the issue when upgrading from a previous version of the chart
that doesn't include the certs secret by ensuring that the certs secret
is created if it doesn't exists.

This method also add more flexibility than charts hooks by ensuring that
the certs won't be rotated each time we upgrade the chart.

Found in https://stackoverflow.com/a/61715870